# Accelerated callbacks for integration routines