--- a/.travis.yml
+++ b/.travis.yml
@@ -297,7 +297,7 @@ before_install:
         conda config --set always_yes yes --set changeps1 no;
         conda update -q conda;
         conda info -a;
-        conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION matplotlib numpy scipy pip;
+        conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION matplotlib numpy scipy pip llvmlite;
         source activate test-environment;
     fi
 
@@ -323,7 +323,6 @@ install:
       pip uninstall -y sympy;
       pip install setuptools;
       python setup.py install;
-      conda install llvmlite;
     elif [[ "${TEST_SAGE}" != "true" ]]; then
       python setup.py install;
     fi