--- a/sympy/printing/llvmjitcode.py
+++ b/sympy/printing/llvmjitcode.py
@@ -10,7 +10,7 @@ import ctypes
 
 from sympy.external import import_module
 from sympy.printing.printer import Printer
-from sympy import S
+from sympy import S, IndexedBase
 from sympy.utilities.decorator import doctest_depends_on
 
 llvmlite = import_module('llvmlite')
@@ -92,6 +92,29 @@ class LLVMJitPrinter(Printer):
                         % type(expr))
 
 
+# Used when parameters are passed by array.  Often used in callbacks to
+# handle a variable number of parameters.
+class LLVMJitCallbackPrinter(LLVMJitPrinter):
+    def __init__(self, *args, **kwargs):
+        super(LLVMJitCallbackPrinter, self).__init__(*args, **kwargs)
+
+    def _print_Indexed(self, expr):
+        array, idx = self.func_arg_map[expr.base]
+        offset = int(expr.indices[0].evalf())
+        array_ptr = self.builder.gep(array, [ll.Constant(ll.IntType(32), offset)])
+        fp_array_ptr = self.builder.bitcast(array_ptr, ll.PointerType(self.fp_type))
+        value = self.builder.load(fp_array_ptr)
+        return value
+
+    def _print_Symbol(self, s):
+        array, idx = self.func_arg_map.get(s, [None, 0])
+        array_ptr = self.builder.gep(array, [ll.Constant(ll.IntType(32), idx)])
+        fp_array_ptr = self.builder.bitcast(array_ptr,
+                                            ll.PointerType(self.fp_type))
+        value = self.builder.load(fp_array_ptr)
+        return value
+
+
 # ensure lifetime of the execution engine persists (else call to compiled
 #   function will seg fault)
 exe_engines = []
@@ -102,19 +125,33 @@ current_link_suffix = 0
 
 
 class LLVMJitCode(object):
-    def __init__(self):
+    def __init__(self, signature):
+        self.signature = signature
         self.fp_type = ll.DoubleType()
         self.module = ll.Module('mod1')
         self.fn = None
-        self.arg_types = []
+        self.llvm_arg_types = []
+        self.llvm_ret_type = self.fp_type
         self.param_dict = {}  # map symbol name to LLVM function argument
         self.link_name = ''
 
+    def _from_ctype(self, ctype):
+        if ctype == ctypes.c_int:
+            return ll.IntType(32)
+        if ctype == ctypes.c_double:
+            return self.fp_type
+        if ctype == ctypes.POINTER(ctypes.c_double):
+            return ll.PointerType(self.fp_type)
+        if ctype == ctypes.c_void_p:
+            return ll.PointerType(ll.IntType(32))
+
+        print("Unhandled ctype = %s" % str(ctype))
+
     def _create_args(self, func_args):
         """Create types for function arguments"""
-        for arg in func_args:
-            arg_type = self.fp_type
-            self.arg_types.append(arg_type)
+        self.llvm_ret_type = self._from_ctype(self.signature.ret_type)
+        self.llvm_arg_types = \
+            [self._from_ctype(a) for a in self.signature.arg_ctypes]
 
     def _create_function_base(self):
         """Create function with name and type signature"""
@@ -124,7 +161,7 @@ class LLVMJitCode(object):
         self.link_name = default_link_name + str(current_link_suffix)
         link_names.add(self.link_name)
 
-        fn_type = ll.FunctionType(self.fp_type, self.arg_types)
+        fn_type = ll.FunctionType(self.llvm_ret_type, self.llvm_arg_types)
         self.fn = ll.Function(self.module, fn_type, name=self.link_name)
 
     def _create_param_dict(self, func_args):
@@ -174,9 +211,58 @@ class LLVMJitCode(object):
         return fptr
 
 
-def _llvm_jit_code(args, expr):
+class LLVMJitCodeCallback(LLVMJitCode):
+    def __init__(self, signature):
+        super(LLVMJitCodeCallback, self).__init__(signature)
+
+    def _create_param_dict(self, func_args):
+        for i, a in enumerate(func_args):
+            if isinstance(a, IndexedBase):
+                self.param_dict[a] = (self.fn.args[i], i)
+                self.fn.args[i].name = str(a)
+            else:
+                self.param_dict[a] = (self.fn.args[self.signature.input_arg],
+                                      i)
+
+    def _create_function(self, expr):
+        """Create function body and return LLVM IR"""
+        bb_entry = self.fn.append_basic_block('entry')
+        builder = ll.IRBuilder(bb_entry)
+
+        lj = LLVMJitCallbackPrinter(self.module, builder, self.fn,
+                                    func_arg_map=self.param_dict)
+
+        ret = lj._print(expr)
+
+        if self.signature.ret_arg:
+            builder.store(ret, self.fn.args[self.signature.ret_arg])
+            builder.ret(ll.Constant(ll.IntType(32), 0))  # return success
+        else:
+            lj.builder.ret(ret)
+
+        strmod = str(self.module)
+        return strmod
+
+
+class CodeSignature(object):
+    def __init__(self, ret_type):
+        self.ret_type = ret_type
+        self.arg_ctypes = []
+
+        # Input argument array element index
+        self.input_arg = 0
+
+        # For the case output value is referenced through a parameter rather
+        # than the return value
+        self.ret_arg = None
+
+
+def _llvm_jit_code(args, expr, signature, callback_type):
     """Create a native code function from a Sympy expression"""
-    jit = LLVMJitCode()
+    if callback_type is None:
+        jit = LLVMJitCode(signature)
+    else:
+        jit = LLVMJitCodeCallback(signature)
 
     jit._create_args(args)
     jit._create_function_base()
@@ -189,8 +275,8 @@ def _llvm_jit_code(args, expr):
     return fptr
 
 
-@doctest_depends_on(modules=('llvmlite',))
-def llvm_callable(args, expr):
+@doctest_depends_on(modules=('llvmlite', 'scipy'))
+def llvm_callable(args, expr, callback_type=None):
     '''Compile function from a Sympy expression
 
     Expressions are evaluated using double precision arithmetic.
@@ -205,6 +291,12 @@ def llvm_callable(args, expr):
         a double precision scalar.
     expr : Expr
         Expression to compile.
+    callback_type : string
+        Create function with signature appropriate to use as a callback.
+        Currently supported:
+           'scipy.integrate'
+           'scipy.integrate.test'
+           'cubature'
 
     Returns
     =======
@@ -220,16 +312,66 @@ def llvm_callable(args, expr):
     3.31000000000000
     >>> e1(1.1)  # Evaluate using JIT-compiled code
     3.3100000000000005
+
+
+    Callbacks for integration functions can be JIT compiled.
+    >>> import sympy.printing.llvmjitcode as jit
+    >>> from sympy.abc import a
+    >>> from sympy import integrate
+    >>> from scipy.integrate import quad
+    >>> e = a*a
+    >>> e1 = jit.llvm_callable([a], e, callback_type='scipy.integrate')
+    >>> integrate(e, (a, 0.0, 2.0))
+    2.66666666666667
+    >>> quad(e1, 0.0, 2.0)[0]
+    2.66666666666667
+
+    The 'cubature' callback is for the Python wrapper around the
+    cubature package ( https://github.com/saullocastro/cubature )
+    and ( http://ab-initio.mit.edu/wiki/index.php/Cubature )
+
+    There are two signatures for the SciPy integration callbacks.
+    The first ('scipy.integrate') is the function to be passed to the
+    integration routine, and will pass the signature checks.
+    The second ('scipy.integrate.test') is only useful for directly calling
+    the function using ctypes variables. It will not pass the signature checks
+    for scipy.integrate.
     '''
 
     if not llvmlite:
         raise ImportError("llvmlite is required for llvmjitcode")
 
-    fptr = _llvm_jit_code(args, expr)
-    arg_ctypes = []
-    for arg in args:
-        arg_ctype = ctypes.c_double
-        arg_ctypes.append(arg_ctype)
+    signature = CodeSignature(ctypes.c_double)
 
-    cfunc = ctypes.CFUNCTYPE(ctypes.c_double, *arg_ctypes)(fptr)
+    arg_ctypes = []
+    if callback_type is None:
+        for arg in args:
+            arg_ctype = ctypes.c_double
+            arg_ctypes.append(arg_ctype)
+    elif callback_type == 'scipy.integrate' or callback_type == 'scipy.integrate.test':
+        arg_ctypes = [ctypes.c_int, ctypes.POINTER(ctypes.c_double)]
+        arg_ctypes_formal = [ctypes.c_int, ctypes.c_double]
+        signature.input_arg = 1
+    elif callback_type == 'cubature':
+        arg_ctypes = [ctypes.c_int,
+                      ctypes.POINTER(ctypes.c_double),
+                      ctypes.c_void_p,
+                      ctypes.c_int,
+                      ctypes.POINTER(ctypes.c_double)
+                      ]
+        signature.ret_type = ctypes.c_int
+        signature.input_arg = 1
+        signature.ret_arg = 4
+    else:
+        print("Unknown callback type: %s" % callback_type)
+        return None
+
+    signature.arg_ctypes = arg_ctypes
+
+    fptr = _llvm_jit_code(args, expr, signature, callback_type)
+
+    if callback_type and callback_type == 'scipy.integrate':
+        arg_ctypes = arg_ctypes_formal
+
+    cfunc = ctypes.CFUNCTYPE(signature.ret_type, *arg_ctypes)(fptr)
     return cfunc