--- a/sympy/printing/tests/test_llvmjit.py
+++ b/sympy/printing/tests/test_llvmjit.py
@@ -55,6 +55,24 @@ def test_two_func():
     assert isclose(jit_res, res)
 
 
+def test_two_sqrt():
+    e = 4.0*sympy.sqrt(a) + sympy.sqrt(b)
+    f = g.llvm_callable([a, b], e)
+    res = float(e.subs({a: 1.5, b: 2.0}).evalf())
+    jit_res = f(1.5, 2.0)
+
+    assert isclose(jit_res, res)
+
+
+def test_two_pow():
+    e = a**1.5 + b**7
+    f = g.llvm_callable([a, b], e)
+    res = float(e.subs({a: 1.5, b: 2.0}).evalf())
+    jit_res = f(1.5, 2.0)
+
+    assert isclose(jit_res, res)
+
+
 def test_callback():
     e = a + 1.2
     f = g.llvm_callable([a], e, callback_type='scipy.integrate.test')