@asmeurer 2016-02-22
Relevant https://stackoverflow.com/questions/35430479/convert-sympy-expressions-to-function-of-numpy-arrays