@asmeurer 2016-02-22
I take it https://github.com/sympy/sympy/pull/10451 should be merged first.