@markdewing 2016-02-24
Thanks for the pointer to the Stack Overflow question. There are two steps that could speed performance
1. Compile to code that accepts and returns numpy arrays.   The evaluation of the function is faster, but there is still the overhead of the scipy wrapper code around the underlying ODE integrator.
2. Modify scipy to accept raw callbacks (as was done for `quad`), and then compile to code for that interface.