@asmeurer 2016-02-26
Is this ready to be reviewed/merged or are more changes coming?