@markdewing 2016-02-26
It should be ready to go. The changes for addressing the stack overflow issue you pointed to are more involved to add here, and would go in a separate PR.