@markdewing 2016-03-07
Here's an example using Sympy plus these integration callbacks to compute the ground state energy of a  hydrogen atom.  It also displays some of the performance of various callbacks/integration packages.

https://github.com/markdewing/next_steps_in_programming/blob/master/examples/integration/Hydogen%20Atom.ipynb